lexer update
