+++ Makefile
@@ -4,9 +4,11 @@ SRC :=	src/main.c \
 		src/tools.c \
 		src/error.c \
 		src/history.c \
-    	src/expander.c \
-    	src/lexer.c \
-		src/lexer_utils.c\
+    	src/lexer/expander.c \
+    	src/lexer/lexer.c \
+		src/lexer/lexer_utils.c \
+		src/lexer/lexer_split.c \
+		src/lexer/lexer_split_utils.c 
 
 OBJ := $(SRC:.c=.o)
