+++ src/lexer/expander.c
@@ -1,4 +1,4 @@
-#include "../includes/minishell.h"
+#include "../../includes/minishell.h"
 
 void	expander(char **list, char **env)
 {
